docs: update example and documentation
