+++ examples/ai-core/package.json
@@ -3,12 +3,12 @@
   "version": "0.0.0",
   "private": true,
   "dependencies": {
-    "@ai-sdk/black-forest-labs": "workspace:*",
     "@ai-sdk/amazon-bedrock": "workspace:*",
     "@ai-sdk/anthropic": "workspace:*",
     "@ai-sdk/assemblyai": "workspace:*",
     "@ai-sdk/azure": "workspace:*",
     "@ai-sdk/baseten": "workspace:*",
+    "@ai-sdk/black-forest-labs": "workspace:*",
     "@ai-sdk/cerebras": "workspace:*",
     "@ai-sdk/cohere": "workspace:*",
     "@ai-sdk/deepgram": "workspace:*",
@@ -22,9 +22,10 @@
     "@ai-sdk/google": "workspace:*",
     "@ai-sdk/google-vertex": "workspace:*",
     "@ai-sdk/groq": "workspace:*",
+    "@ai-sdk/huggingface": "workspace:*",
+    "@ai-sdk/hume": "workspace:*",
     "@ai-sdk/lmnt": "workspace:*",
     "@ai-sdk/luma": "workspace:*",
-    "@ai-sdk/hume": "workspace:*",
     "@ai-sdk/mcp": "workspace:*",
     "@ai-sdk/mistral": "workspace:*",
     "@ai-sdk/openai": "3.0.0-beta.107",
@@ -37,14 +38,15 @@
     "@ai-sdk/valibot": "workspace:*",
     "@ai-sdk/vectorstores": "workspace:*",
     "@ai-sdk/vercel": "workspace:*",
-    "@vectorstores/core": "^0.1.3",
     "@ai-sdk/xai": "workspace:*",
-    "@ai-sdk/huggingface": "workspace:*",
     "@google/generative-ai": "0.21.0",
     "@opentelemetry/auto-instrumentations-node": "0.54.0",
     "@opentelemetry/sdk-node": "0.54.2",
     "@opentelemetry/sdk-trace-node": "1.28.0",
+    "@standard-schema/spec": "1.1.0",
     "@valibot/to-json-schema": "^1.3.0",
+    "@vectorstores/core": "^0.1.3",
+    "@vectorstores/readers": "^0.1.3",
     "ai": "workspace:*",
     "arktype": "2.1.28",
     "dotenv": "16.4.5",
@@ -52,7 +54,6 @@
     "image-type": "^5.2.0",
     "mathjs": "14.0.0",
     "sharp": "^0.33.5",
-    "@standard-schema/spec": "1.1.0",
     "terminal-image": "^2.0.0",
     "valibot": "1.1.0",
     "zod": "3.25.76"
